feat: Add skeleton loader for app initialization
Introduces a skeleton loader component to provide a better user experience during the initial loading of the application data. This component displays a series of placeholder elements, mimicking the structure of the dashboard, while the actual data is being fetched.

A new CSS animation `pulse-bg` has been added to `index.html` to facilitate the pulsing effect of the skeleton elements. The `isDbLoaded` state has been added to `AppContext` to control the visibility of the skeleton loader versus the main application content.