feat(Button): Set default button type to 'button'
This change sets the default `type` attribute for the Button component to `button`. This prevents accidental form submissions when a Button is used outside of a form or when its primary purpose is not to submit a form.

Also, prevents default form submission when opening the print modal for products.